+++ postgres/src/execute/async_impl.rs
@@ -1,19 +1,14 @@
-use core::{
-    future::{ready, Future, Ready},
-    pin::Pin,
-    task::{Context, Poll},
-};
+use core::future::{ready, Future, Ready};
 
 use crate::{
     driver::codec::AsParams,
     error::Error,
     prepare::Prepare,
-    query::{Query, RowAffected, RowSimpleStream, RowStream, RowStreamGuarded},
+    query::{Query, RowSimpleStream, RowStream, RowStreamGuarded},
     statement::{
         Statement, StatementCreate, StatementGuarded, StatementNamed, StatementQuery, StatementUnnamedBind,
         StatementUnnamedQuery,
     },
-    BoxedFuture,
 };
 
 use super::Execute;
@@ -22,16 +17,17 @@ impl<'s, C> Execute<&C> for &'s Statement
 where
     C: Query,
 {
-    type ExecuteOutput = ResultFuture<RowAffected>;
-    type QueryOutput = Ready<Result<RowStream<'s>, Error>>;
+    type ExecuteOutput = Result<u64, Error>;
+    type QueryOutput = Result<RowStream<'s>, Error>;
 
     #[inline]
-    fn execute(self, cli: &C) -> Self::ExecuteOutput {
-        cli._query(self).map(RowAffected::from).into()
+    fn execute(self, cli: &C) -> impl Future<Output = Self::ExecuteOutput> + 'static {
+        let res = cli._query(self).map(|res| res.res);
+        async { res?.try_into_row_affected().await }
     }
 
     #[inline]
-    fn query(self, cli: &C) -> Self::QueryOutput {
+    fn query(self, cli: &C) -> Ready<Self::QueryOutput> {
         ready(cli._query(self))
     }
 }
@@ -40,39 +36,18 @@ impl<C> Execute<&C> for &str
 where
     C: Query,
 {
-    type ExecuteOutput = ResultFuture<RowAffected>;
-    type QueryOutput = Ready<Result<RowSimpleStream, Error>>;
+    type ExecuteOutput = Result<u64, Error>;
+    type QueryOutput = Result<RowSimpleStream, Error>;
 
     #[inline]
-    fn execute(self, cli: &C) -> Self::ExecuteOutput {
-        cli._query(self).map(RowAffected::from).into()
+    fn execute(self, cli: &C) -> impl Future<Output = Self::ExecuteOutput> + 'static {
+        let res = cli._query(self).map(|res| res.res);
+        async { res?.try_into_row_affected().await }
     }
 
     #[inline]
-    fn query(self, cli: &C) -> Self::QueryOutput {
+    fn query(self, cli: &C) -> Ready<Self::QueryOutput> {
         ready(cli._query(self))
-    }
-}
-
-type IntoGuardedFuture<'c, C> = IntoGuarded<'c, BoxedFuture<'c, Result<Statement, Error>>, C>;
-
-pub struct IntoGuarded<'a, F, C> {
-    fut: F,
-    cli: &'a C,
-}
-
-impl<'a, F, C> Future for IntoGuarded<'a, F, C>
-where
-    F: Future<Output = Result<Statement, Error>> + Unpin,
-    C: Query,
-{
-    type Output = Result<StatementGuarded<'a, C>, Error>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let this = self.get_mut();
-        Pin::new(&mut this.fut)
-            .poll(cx)
-            .map_ok(|stmt| stmt.into_guarded(this.cli))
     }
 }
 
@@ -80,18 +55,18 @@ impl<'c, C> Execute<&'c C> for StatementNamed<'_>
 where
     C: Prepare,
 {
-    type ExecuteOutput = ResultFuture<IntoGuardedFuture<'c, C>>;
+    type ExecuteOutput = Result<StatementGuarded<'c, C>, Error>;
     type QueryOutput = Self::ExecuteOutput;
 
     #[inline]
-    fn execute(self, cli: &'c C) -> Self::ExecuteOutput {
-        cli._query(StatementCreate::from((self, cli)))
-            .map(|fut| IntoGuarded { fut, cli })
-            .into()
+    async fn execute(self, cli: &'c C) -> Self::ExecuteOutput {
+        cli._query(StatementCreate::from((self, cli)))?
+            .await
+            .map(|stmt| stmt.into_guarded(cli))
     }
 
     #[inline]
-    fn query(self, cli: &'c C) -> Self::QueryOutput {
+    fn query(self, cli: &'c C) -> impl Future<Output = Self::QueryOutput> {
         self.execute(cli)
     }
 }
@@ -101,16 +76,17 @@ where
     C: Query,
     P: AsParams,
 {
-    type ExecuteOutput = ResultFuture<RowAffected>;
-    type QueryOutput = Ready<Result<RowStream<'s>, Error>>;
+    type ExecuteOutput = Result<u64, Error>;
+    type QueryOutput = Result<RowStream<'s>, Error>;
 
     #[inline]
-    fn execute(self, cli: &C) -> Self::ExecuteOutput {
-        cli._query(self).map(RowAffected::from).into()
+    fn execute(self, cli: &C) -> impl Future<Output = Self::ExecuteOutput> + 'static {
+        let res = cli._query(self).map(|res| res.res);
+        async { res?.try_into_row_affected().await }
     }
 
     #[inline]
-    fn query(self, cli: &C) -> Self::QueryOutput {
+    fn query(self, cli: &C) -> Ready<Self::QueryOutput> {
         ready(cli._query(self))
     }
 }
@@ -120,18 +96,19 @@ where
     C: Prepare,
     P: AsParams,
 {
-    type ExecuteOutput = ResultFuture<RowAffected>;
-    type QueryOutput = Ready<Result<RowStreamGuarded<'c, C>, Error>>;
+    type ExecuteOutput = Result<u64, Error>;
+    type QueryOutput = Result<RowStreamGuarded<'c, C>, Error>;
 
     #[inline]
-    fn execute(self, cli: &C) -> Self::ExecuteOutput {
-        cli._query(StatementUnnamedQuery::from((self, cli)))
-            .map(RowAffected::from)
-            .into()
+    async fn execute(self, cli: &C) -> Self::ExecuteOutput {
+        cli._query(StatementUnnamedQuery::from((self, cli)))?
+            .res
+            .try_into_row_affected()
+            .await
     }
 
     #[inline]
-    fn query(self, cli: &'c C) -> Self::QueryOutput {
+    fn query(self, cli: &'c C) -> Ready<Self::QueryOutput> {
         ready(cli._query(StatementUnnamedQuery::from((self, cli))))
     }
 }
@@ -140,52 +117,26 @@ impl<'c, C> Execute<&'c C> for &std::path::Path
 where
     C: Query + Sync,
 {
-    type ExecuteOutput = BoxedFuture<'c, Result<u64, Error>>;
-    type QueryOutput = BoxedFuture<'c, Result<RowSimpleStream, Error>>;
+    type ExecuteOutput = Result<u64, Error>;
+    type QueryOutput = Result<RowSimpleStream, Error>;
 
     #[inline]
-    fn execute(self, cli: &'c C) -> Self::ExecuteOutput {
+    async fn execute(self, cli: &'c C) -> Self::ExecuteOutput {
         let path = self.to_path_buf();
-        Box::pin(async move {
-            tokio::task::spawn_blocking(|| std::fs::read_to_string(path))
-                .await
-                .unwrap()?
-                .execute(cli)
-                .await
-        })
+        tokio::task::spawn_blocking(|| std::fs::read_to_string(path))
+            .await
+            .unwrap()?
+            .execute(cli)
+            .await
     }
 
     #[inline]
-    fn query(self, cli: &'c C) -> Self::QueryOutput {
+    async fn query(self, cli: &'c C) -> Self::QueryOutput {
         let path = self.to_path_buf();
-        Box::pin(async move {
-            tokio::task::spawn_blocking(|| std::fs::read_to_string(path))
-                .await
-                .unwrap()?
-                .query(cli)
-                .await
-        })
-    }
-}
-
-pub struct ResultFuture<F>(Result<F, Option<Error>>);
-
-impl<F> From<Result<F, Error>> for ResultFuture<F> {
-    fn from(res: Result<F, Error>) -> Self {
-        Self(res.map_err(Some))
-    }
-}
-
-impl<F, T> Future for ResultFuture<F>
-where
-    F: Future<Output = Result<T, Error>> + Unpin,
-{
-    type Output = F::Output;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        match self.get_mut().0 {
-            Ok(ref mut res) => Pin::new(res).poll(cx),
-            Err(ref mut e) => Poll::Ready(Err(e.take().unwrap())),
-        }
+        tokio::task::spawn_blocking(|| std::fs::read_to_string(path))
+            .await
+            .unwrap()?
+            .query(cli)
+            .await
     }
 }